others/Makefile: stealthMode and optware no longer to be installed.
- stealthMode was originally written for mipsel routers. I heard of problems of Yahoo! API, but not sure if they were rectified, but it doesn't matter.
stealthMode cannot be used with ARM routers because of blink binary. It is easy enough to kill blink, but if users do any change that causes wireless to be restarted, blink will also be restarted, and will nullify the effects of this script.

- optware is not updated since 2014.
And we have Entware for years, which is well maintained and supported as of 2017.