Preliminary support for 2nd 5Ghz LED on R8000.
Added preliminary support of the 2nd 5Ghz led for led command. It's supposed to honor the GPIO assignment for the device (gpio disable|enable 16) but doesn't work at the moment. It's the same deal with USB LED's that won't turn on|off with led commands.
Logic for blink startup was extended because it is in a for loop and we have three wireless interfaces. If rc/network.c code was left as is, blink would start referencing the 3rd interface but still referencing 1st 5Ghz LED.
There's a kludge to start 2nd 5Ghz LED directly via a gpio command, instead of led command. There is no point changing at the moment because it won't work.

Feedback welcome.