Revert "R7000 / R6400: WLAN LED cleanup - revert commit https://bitbucket.org/kille72/freshtomato-arm/commits/e990a317a838ff040cc0e486fb2c2866aea66fdf (because WLAN LED is already turned on/off with blink) - use blink for WLAN LED (2.4 GHz and 5 GHz)"
In R7000/R6400, the actual LED which LED_AOSS is in control. informs the overall wi-fi radio status (called GPIO_WIFI_SUMMARY_LED in Genie sources).
If wi-fi is completely off, that LED should turn off as well. This attempts to replicate Genie's behavior with that specific LED.

The other commands for LED_WLAN and LED_5G (and LED_52G on SDK7) can be removed, but for now, I need to revert this one.

This reverts commit e6f8dfffb56882a870312454382f8a2bce4d76fa.