pppd: fixes from upstream
- abort on errors in subdir builds: https://github.com/paulusmack/ppp/commit/5e90783d11a59268e05f4cfb29ce2343b13e8ab2

The current recursive loops do not check the exit status of make
in subdirs which leads to `make` passing even when a subdir failed
to compile or install.

URL: https://bugs.gentoo.org/334727
Signed-off-by: Martin von Gagern <Martin.vGagern@gmx.net>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>

- pppd: Fix sign-extension when displaying bytes in octal: https://github.com/paulusmack/ppp/commit/5e8c3cb256a7e86e3572a82a75d51c6850efdbdc

print_string() displays characters as \\%.03o but without first
casting it from "char" to "unsigned char" so it gets sign-extended
to an int. This causes output like \37777777630 instead of \230.

Signed-off-by: Philip A. Prindeville <philipp@redfish-solutions.com>

- Suppress false error message on PPPoE disconnect: https://github.com/paulusmack/ppp/commit/219aac3b53d0827549377f1bfe22853ee52d4405

Once the kernel handles PPPoE PADTs correctly[1], a PADT triggered
disconnect will result in EALREADY when pppd tries to clear the session ID.

Simply ignore the error if, and only if, the error is EALREADY

[1] https://patchwork.ozlabs.org/patch/444717/

Signed-off-by: Simon Farnsworth <simon@farnz.org.uk>

- Send PADT on PPPoE disconnect: https://github.com/paulusmack/ppp/commit/cd2c14f998c57bbe6a01dc5854f2763c0d7f31fb

Once we've terminated the PPP session, there is no chance of a PPP layer
disconnect. Some PPPoE relays don't detect the PPP session going down, and
depend on a long timeout or a PPPoE PADT to terminate the session.

Send a PADT on disconnect to work around these buggy relays.

Signed-off-by: Simon Farnsworth <simon@farnz.org.uk>

- pppd: ipxcp: Prevent buffer overrun on remote router name: https://github.com/paulusmack/ppp/commit/fe149de624f96629a7f46732055d8f718c74b856

This fixes an if condition to prevent a possible 1-byte overrun
on ipxcp_hisoptions[0].name.

Reported-by: "Sabas Rosales, Blanca E" <blanca.e.sabas.rosales@intel.com>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>

- pppd: Fix ccp_cilen calculated size if both deflate_correct and deflate_draft are enabled: https://github.com/paulusmack/ppp/commit/094cb8ae4c61db225e67fedadb4964f846dd0c27

This fixes a bug where ccp_cilen() will return 4 bytes less than
necessary for the addci buffer if both deflate_correct and
deflate_draft are enabled.

Signed-off-by: Sylvain Rochet <gradator@gradator.net>
Signed-off-by: Paul Mackerras <paulus@samba.org>

- pppd: Remove unused declaration of ttyname: https://github.com/paulusmack/ppp/commit/66ae569f1c6fb12df8a463c8beb9afc0c7e5f7f0

Signed-off-by: George Burgess IV <george@gbiv.net>

- pppd: Fix printing call in print_option(): https://github.com/paulusmack/ppp/commit/3b409a8e299446e3118743ff2508c54330011eeb

print_option() was in this case passing p/opt->addr2 as the format string
instead of the string to be printed (as a quoted string). That could lead
to a nasty crash.

Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>

- pppd: Don't try to free(NULL - 1): https://github.com/paulusmack/ppp/commit/b5a2fd5ca2cd00b564d829c352d825902fba13a0

A logic bug in update_script_environment() means that it can call
remove_script_env() even when the variable being removed is not
present in the script_env array.  The result of that is that
remove_script_env() will call free() with argument NULL - 1.
To fix this, we avoid calling remove_script_env() in this case.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>

- Change include from sys/errno.h to errno.h: https://github.com/paulusmack/ppp/commit/56a4760236ee0a606075404c5704801feffa646c

According to POSIX, the canonical location for errno.h is on the top level.

Signed-off-by: Stefan Nickl <Stefan.Nickl@gmail.com>

- pppoe: include netinet/in.h before linux/in.h: https://github.com/paulusmack/ppp/commit/50a2997b256e0e0ef7a46fae133f56f60fce539c

This fixes builds with newer kernels. Basically, <netinet/in.h> needs to be
included before <linux/in.h> otherwise the earlier, unaware of the latter,
tries to redefine symbols and structures. Also, <linux/if_pppox.h> doesn't work
alone anymore, since it pulls the headers in the wrong order, so we better
include <netinet/in.h> early.